+++ src/Stack/Solver.hs
@@ -28,7 +28,7 @@ import           Data.Text.Encoding          (decodeUtf8, encodeUtf8)
 import qualified Data.Yaml                   as Yaml
 import           Network.HTTP.Client.Conduit (HasHttpManager)
 import           Path
-import           Path.IO                     (withCanonicalizedSystemTempDirectory)
+import           Path.IO                     (parseRelAsAbsDir)
 import           Prelude
 import           Stack.BuildPlan
 import           Stack.Types
@@ -36,6 +36,7 @@ import           System.Directory            (copyFile,
                                               createDirectoryIfMissing,
                                               getTemporaryDirectory)
 import qualified System.FilePath             as FP
+import           System.IO.Temp              (withSystemTempDirectory)
 import           System.Process.Read
 
 cabalSolver :: (MonadIO m, MonadLogger m, MonadMask m, MonadBaseControl IO m, MonadReader env m, HasConfig env)
@@ -44,7 +45,7 @@ cabalSolver :: (MonadIO m, MonadLogger m, MonadMask m, MonadBaseControl IO m, Mo
             -> Map PackageName Version -- ^ constraints
             -> [String] -- ^ additional arguments
             -> m (CompilerVersion, Map PackageName (Version, Map FlagName Bool))
-cabalSolver wc cabalfps constraints cabalArgs = withCanonicalizedSystemTempDirectory "cabal-solver" $ \dir -> do
+cabalSolver wc cabalfps constraints cabalArgs = withSystemTempDirectory "cabal-solver" $ \dir -> do
     configLines <- getCabalConfig dir constraints
     let configFile = dir FP.</> "cabal.config"
     liftIO $ S.writeFile configFile $ encodeUtf8 $ T.unlines configLines
@@ -58,7 +59,7 @@ cabalSolver wc cabalfps constraints cabalArgs = withCanonicalizedSystemTempDirec
     --
     -- In theory we could use --ignore-sandbox, but not all versions of cabal
     -- support it.
-    tmpdir <- liftIO getTemporaryDirectory >>= parseAbsDir
+    tmpdir <- liftIO getTemporaryDirectory >>= parseRelAsAbsDir
 
     let args = ("--config-file=" ++ configFile)
              : "install"