+++ ChangeLog.md
@@ -22,6 +22,7 @@ Bug fixes:
   [#1052](https://github.com/commercialhaskell/stack/issues/1052)
 * Fix quoting of output failed command line arguments
 * Mark executable-only packages as installed when copied from cache [#1043](https://github.com/commercialhaskell/stack/pull/1043)
+* Canonicalize temporary directory paths [#1047](https://github.com/commercialhaskell/stack/pull/1047)
 
 ## 0.1.5.0